docs: sendpay_success timestamps are u32
json_add_payment_fields() emits created_at and completed_at with
json_add_u32, matching struct wallet_payment:

  lightningd/pay.c:158
    json_add_u32(response, "created_at", t->timestamp);
  lightningd/pay.c:160
    json_add_u32(response, "completed_at", *t->completed_at);
  wallet/wallet.h:363-364
    u32 timestamp; u32 *completed_at;

Changelog-None

+++ cln-rpc/src/notifications.rs
@@ -812,7 +812,7 @@ pub struct SendPaySuccessNotification {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub bolt12: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub completed_at: Option<u64>,
+    pub completed_at: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub description: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -830,7 +830,7 @@ pub struct SendPaySuccessNotification {
     // Path `sendpay_success.status`
     pub status: SendpaySuccessStatus,
     pub amount_sent_msat: Amount,
-    pub created_at: u64,
+    pub created_at: u32,
     pub created_index: u64,
     pub groupid: u64,
     pub id: u64,
